Moved files to packages Initial commit for ConnectionManager
