+++ app/src/main/java/distudios/at/carcassonne/networking/lobby/OnWifiP2pDeviceActionEventListener.java
@@ -1,9 +1,6 @@
-package distudios.at.carcassonne.networking;
+package distudios.at.carcassonne.networking.lobby;
 
 import android.net.wifi.p2p.WifiP2pDevice;
-import android.widget.Button;
-
-import distudios.at.carcassonne.gui.MyDeviceItemRecyclerViewAdapter;
 
 public interface OnWifiP2pDeviceActionEventListener {
     void onAction(WifiP2pDevice device, MyDeviceItemRecyclerViewAdapter.ViewHolder holder);
