+++ app/src/main/res/layout/fragment_deviceitem_list.xml
@@ -9,5 +9,5 @@
     android:layout_marginLeft="16dp"
     android:layout_marginRight="16dp"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".gui.DeviceListFragment"
+    tools:context=".networking.lobby.DeviceListFragment"
     tools:listitem="@layout/fragment_deviceitem" />
